refactor: console.log 제거 및 api 파일 정리
