+++ src/api/edit.ts
@@ -1,7 +1,21 @@
-// 문제 수정 요청
-
-import { authApiClient } from "../api/api-client";
+import { authApiClient } from "@/api/api-client";
 import usePocketStore from "@/store/pocket";
+
+export const getEdit = async (questionCustomId: number) => {
+  try {
+    const response = await authApiClient.get(
+      `/api/question/${questionCustomId}`,
+    );
+    if (response.status === 200) {
+      return response.data.data; // 문제 데이터를 반환
+    } else {
+      throw new Error("문제 데이터를 가져오지 못했습니다.");
+    }
+  } catch (error) {
+    console.error("문제 가져오는 중 오류 발생:", error);
+    throw new Error("문제 데이터를 가져오는 데 실패했습니다.");
+  }
+};
 
 export const postEdit = async (
   title: string,
@@ -34,7 +48,6 @@ export const postEdit = async (
     );
 
     if (response.status === 200) {
-      console.log("문제 수정 성공:", response.data);
       return response.data.data; // 수정된 데이터 반환
     } else {
       throw new Error("문제 수정을 완료하지 못했습니다.");
