+++ src/api/form.ts
@@ -7,7 +7,6 @@ import usePocketStore from "@/store/pocket";
 export const fetchLuckyPouches = async (page: number = 0): Promise<any> => {
   try {
     const response = await authApiClient.get(`/api/luckyPouch?page=${page}`);
-    console.log("복주머니 불러오기 성공:", response.data);
     return response.data.data;
   } catch (error) {
     console.error("복주머니 불러오기 실패:", error);
@@ -25,7 +24,6 @@ export const fetchRandomQuestion = async (): Promise<{
 }> => {
   try {
     const response = await authApiClient.get("/api/pouch/question");
-    console.log(response);
     return response.data.data;
   } catch (error) {
     //console.error("랜덤 질문 가져오기 실패:", error);
@@ -44,18 +42,18 @@ export const submitCustomQuestion = async (
   content: string | null,
   //paper: string | null
 ): Promise<{ questionId: any }> => {
-  console.log({
-    title,
-    select1: answers[0],
-    select2: answers[1],
-    select3: answers[2],
-    select4: answers[3],
-    answer: correctAnswer,
-    card,
-    domain,
-    content,
-    //paper,
-  });
+  // console.log({
+  //   title,
+  //   select1: answers[0],
+  //   select2: answers[1],
+  //   select3: answers[2],
+  //   select4: answers[3],
+  //   answer: correctAnswer,
+  //   card,
+  //   domain,
+  //   content,
+  //   //paper,
+  // });
   try {
     const response = await authApiClient.post("/api/question", {
       title,
@@ -70,7 +68,6 @@ export const submitCustomQuestion = async (
       //paper,
     });
 
-    console.log("응답데이터");
     // 응답 처리
     if (response.status === 200) {
       const questionId = response.data.data.id;
@@ -78,7 +75,7 @@ export const submitCustomQuestion = async (
       setQuestionId(questionId);
       // 응답할때 questionCustomId 받아오면 이걸 zustand 에 일단 저장해놓음
       // -> url 만들 때 쓸거니까
-      console.log("POST 성공:  quesitonCustomId:", questionId);
+
       return { questionId };
     } else {
       throw new Error(" 제대로 받아오지 못함");
